PURE data

+++ src/preprocess.py
@@ -14,7 +14,6 @@ class Preprocessor:
     def __init__(self):
         self.vocab_file = os.path.join(VOCAB_DIR, "vocabulary.txt")
         self.vocab_with_wiki = os.path.join(VOCAB_DIR, "vocabulary_wiki.txt")
-        self.w2v_raw_dir = os.path.join(RAW_DIR, "SE_Books_txt")
 
     def __wiki_context_extract(self, word):
         try:
@@ -29,6 +28,8 @@ class Preprocessor:
 
     def add_wiki_content(self):
         count = 0
+        if os.path.isfile(self.vocab_with_wiki):
+            return
         with open(self.vocab_file, encoding="utf8") as fin, open(self.vocab_with_wiki, 'w', encoding="utf8") as fout:
             for line in fin:
                 word = line.strip("\n\t\r")
@@ -42,22 +43,29 @@ class Preprocessor:
     def read_all_files(self):
         """
         Read all files in a directory and make the content in each file to be lowercase. All files will be merged to a
-        single string
+        single string. The file path can be too long for windows machine.
         :param dir_path: The path to the directory
         :return: A string representing all files
         """
+
         self.add_wiki_content()
-        file_names = os.listdir(self.w2v_raw_dir)
-        file_paths = [os.path.join(self.w2v_raw_dir, file_name) for file_name in file_names]
+        file_paths = []
+        for raw_dir in RAW_DIR_LIST:
+            file_names = os.listdir(raw_dir)
+            file_paths.extend([os.path.join(raw_dir, file_name) for file_name in file_names])
         file_paths.append(self.vocab_with_wiki)
+
         raw_txt = ''
         for file_path in file_paths:
             if not file_path.endswith("txt"):
                 continue
             print("Processing " + file_path)
-            with open(file_path, 'r', encoding='utf8') as input:
-                for line in input:
-                    raw_txt += line.lower()
+            try:
+                with open(file_path, 'r', encoding='utf8') as input:
+                    for line in input:
+                        raw_txt += line.lower()
+            except Exception as e:
+                pass
         return raw_txt
 
     def tokenize(self, raw_txt):
@@ -121,12 +129,31 @@ class Preprocessor:
         return re_stop_sentences
 
     def write_to_file(self, output_path, sentences):
+        print("Writing to disk ...")
         with open(output_path, 'w', encoding='utf8') as outfile:
             for line in sentences:
                 line = ' '.join(line).strip(' \n') + '\n'
                 if len(line) >= 10:
                     outfile.write(line)
 
+
+'''
+class PUREDataExtractor:
+    def __init__(self):
+        xmls = [os.path.join(PURE_REQ_DIR, fname) for fname in os.listdir(PURE_REQ_DIR) if fname.endswith(".xml")]
+        for xml_path in xmls:
+            print("processing " + xml_path)
+            tree = ET.parse(xml_path)
+            txt = ""
+            for element in tree.iter():
+                if element.tag.endswith("title") or element.tag.endswith("text_body"):
+                    if (element.text == None):
+                        continue
+                    txt += element.text.strip("\n\t\r") + "\n"
+            txt_path = xml_path[:-len(".xml")] + ".txt"
+            with open(txt_path, 'w', encoding='utf8') as of:
+                of.write(txt)
+'''
 
 if __name__ == "__main__":
     print("Start ...")
